Hack up cmdq, a way to run bash commands locally
Not quite bugless, but pretty close to what I wanted, which is to dumpa
bunch of commands on another terminal serially.

If you queue before you pull, the elements tend to land in any order,
but if you're reading it does a decent job of locking fairly.

Domain sockets don't quite work as intended, but that should solve both
problems neatly, once and for all.